@@ -1,22 +1,28 @@
 #ifndef ACTOR_H
 #define ACTOR_H
 
-#include <string>
+#include "main.h"
 
 class Actor {
 public :
     int x; // Actor x position on console
     int y; // Actor y position on console
     int ch; // Actor ASCII code
-    int maxHP = 100; //
-    int curHP = 100;
+    bool blocks;
+    const char *name;
+    //int maxHP = 100;
+    //int curHP = 100;
     bool isPlayer = false;
+    Destructible *destructible;
+    Attacker *attacker;
+    Ai *ai; // something self-updating
 
-    std::string name; //character name
     TCODColor col; // Actor color
 
-    Actor (int x, int y, int ch, const TCODColor &col, std::string name);
-    ~Actor();
+    Actor (int x, int y, int ch, const TCODColor &col, const char *name);
+    ~Actor() = default;
+    //Actor ( Actor && ) = default;
+    //Actor ( const Actor & ) = default;
 
     void render();
     void update();
